pyproject.toml
[project]
name = "smartdoc-ai"
version = "0.1.0"
description = "基于 Qwen2.5 的智能文档转换与差异比对系统"
readme = "README.md"
requires-python = ">=3.10"
# 核心依赖列表
dependencies = [
    "pymupdf>=1.23.0",        # PDF 基础解析
    "pdfplumber>=0.10.0",     # PDF 表格高精度提取
    "python-docx>=1.1.0",     # Word 文档处理
    "diff-match-patch>=20230711", # 语义比对算法
    "llama-cpp-python>=0.2.0", # 本地 GGUF 推理引擎
    "rich>=13.0.0",           # 黑窗口 UI 美化
    "tabulate>=0.9.0",        # 表格格式化辅助
]

[tool.uv]
# 确保使用本地编译插件以支持 CUDA
native-tls = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"